Refactor for fast read, better exposure control
