+++ setup.py
@@ -1,15 +1,10 @@
 #!/usr/bin/env python
-"""
-On Raspberry Pi:
-
-apt install python3-numpy
-"""
-from setuptools import setup
+from setuptools import setup,find_packages
 
 setup(name='pibayer',
-      packages=['pibayer'],
+      packages=find_packages(),
       author='Michael Hirsch, Ph.D.',
-      version='0.5.1',
+      version='0.6.0',
 	  description='Acquire raw Bayer-masked image from Raspberry Pi camera and write image stack to HDF5 or TIFF.',
 	  long_description=open('README.rst').read(),
 	  url='https://github.com/scivision/raspicam-raw-bayer',
@@ -23,5 +18,5 @@ setup(name='pibayer',
       install_requires=['picamera','numpy','nose'],
       extras_require={'io':['tifffile==0.6','h5py','netcdf4','xarray'],
                       'plot':['matplotlib'],},
-      python_requires='>=3.6',
+      python_requires='>=3.5',
 	  )
